Fixes upstream deployments
There was an issue with patching the overcloud where the patch binary is
missing, making it impossible to apply patches.  This change install
patch now on the image.

Also, although deployments were successful, storage was not working.
This is because by default upstream does not apply Ceph OSDs to compute
nodes for hyperconverged Ceph, but we use this as our standard
deployment in Apex.  This patch inserts CephOSD into the default Compute
role.  Note: we normally override role's services in regular Apex
deployments so we do not hit this issue there.

Change-Id: I5bddda4784dc00148395863ae0990343a4159602
Signed-off-by: Tim Rozet <trozet@redhat.com>