Enables br-ex on controllers in SFC scenario
ODL's current behavior is to select a node to serve as the routing/L3
node in a deployment.  The selection is random, and it could end up
choosing a control node.  Therefore the control node needs to have
external network connectivity so adding br-ex to the controller which
will enable this via default ProviderMappings.

JIRA: APEX-558

Change-Id: Ie3044e5711ba62d73b83d69cde369e7452958844
Signed-off-by: Tim Rozet <trozet@redhat.com>