Fixes including default SDN env file in deploy cmd
The latest method to deploy opendaylight features upstream is to include
the default SDN file and then add-on the feature env file which should
override any default settings in the previous SDN env file.  For
example for bgpvpn:

openstack overcloud deploy -e <tht_dir>/neutron-opendaylight.yaml -e \
<tht_dir>/neutron-bgpvpn-opendaylight.yaml

This change documents the recursive function used to find the SDN env
files and includes the default for the SDN controller.

JIRA: APEX-555

Change-Id: I4270932ca4fbc21ea54e965d9d6491424f4463e2
Signed-off-by: Tim Rozet <trozet@redhat.com>