Update patch set 2
Patch Set 2: Code-Review-1

The patches section in the deploy settings are only valid for upstream deployments.  In this case we want the fix for Fraser (when we build our own RPMs, etc).  So you need to just update ci/PR_revision log in this patch.  Then you need to have an apex-tht patch with your cherry pick there, and reference that change-id in this patch's commit msg.

Patch-set: 2
Label: Code-Review=-1