Fixes stale undercloud delorean repos
When we build artifacts for release, we try to freeze the disk images
and artifacts we use for deployment.  This includes the delorean repo
which provides OpenStack packages.  This repo however expires after a
couple months and then if any packages were missing in the Undercloud
released artifact, they will attempt to be downloaded during install
and fail.

This patch adds some logic to detect if there is internet connectivity,
and if so then download and update the delorean repo on the undercloud.

JIRA: APEX-565

Change-Id: I58c28437fb5c5b179033c939377fd97aefbf427c
Signed-off-by: Tim Rozet <trozet@redhat.com>