Fixes ensuring VBMCs are actually running
We currently start VBMCs using CLI due to issues with the VBMC python
lib.  However when we start them, there is no check to make sure they
are actually changed to 'running' status.  This patch adds logic to
check (up to 5 times) to ensure each VBMC is running or raises an error.
Note this is for virtual deployments only.

JIRA: APEX-527

Change-Id: Iab7ee3b76292d6fc547f18c83f23c04205e9bb2e
Signed-off-by: Tim Rozet <trozet@redhat.com>