Fixes ceph key import failures
There is an issue with HA deployments where sometimes key imports fail
for Ceph which seem to occur around 50% of the time.  When logging in
after a failure, the key import seems to work which indicates it may be
a race condition.  Perhaps it is possible that Ceph Mon is coming up too
slowly and not all the way up when the import key exec is called.

This patch adds a retries for a minute to try to import the key.

JIRA: APEX-563

Change-Id: Ie8cfeb4803f6bed95f9e612eeb37c5cdf2d76617
Signed-off-by: Tim Rozet <trozet@redhat.com>