Update patch set 1
Patch Set 1:

So the reason the node is going offline is during initial deployment...when the jumphost nic files get reconfigured and linux networking is restarted, the node stops responding to ipmi or its external network.  The ip assigned to to external is:
IPADDR="10.10.110.20"
GATEWAY="10.10.110.1"
PREFIX="24"

According to the wiki, the ipmi ip is 10.10.110.10.  I'm not sure if there is some kind of conflict or why ipmi is not working.

Other than this, the fix appears to be working.  The Network Deployment failed in overcloud because the network settings had the wrong type "interface" set for the external network when it should be "ovs_bridge" phys_type

Patch-set: 1