Default ctlplane to native VLAN
We were not setting any vlan type value for the admin network.  This
caused an issue when deploying and trying to use a single collapsed
network (no net isolation).  The issue occurred when trying to create an
external neutron network.  We happen to check if the phys type is flat
or vlan using this attribute to decide what kind of phys type to use for
the neutron network.

JIRA: APEX-606

Change-Id: I4e24dd5e8b99cef920b8203b820a77d0021631cc
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 42a9303a21a46f9e9b1997ab4c0beceb6b0bf438)