Adds symlink for "master" openstack
We currently specify "nofeature" for master OS and OS branch names (like
"queens") for other files.  This patch adds a symlink to create another
config file.  This file is not included in the RPM distribution, it is
simply used in CI for CSIT promotion where we create a scenario name
based on openstack version:

https://github.com/opnfv/releng/blob/master/jjb/apex/apex.yaml.j2#L687

Change-Id: I7b7873fbbc2c4d386206f24eea1c94aa0309913f
Signed-off-by: Tim Rozet <trozet@redhat.com>