Update patch set 1
Patch Set 1:

not sure what is going on with odl, but it is failing to cluster correctly in the verify job.  However the ODL package version is the same we have been using for a while now, and we have never seen this issue.  This patch has nothing to do with ODL and we are moving the verify job to use nosdn:
https://gerrit.opnfv.org/gerrit/#/c/58059/

recheck

Patch-set: 1
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0