Update patch set 1
Patch Set 1:

Yeah just using this now to see if our apex/functest CI is fixed (tempest cases were failing).  But you are right if the fix is merged we ignore it, but in the case where the code has not yet been promoted into current-tripleo-rdo...we have no way of knowing that.  I'm not sure what the easiest way to do that is.  Any ideas?

Patch-set: 1