Update patch set 1
Patch Set 1:

> Yeah just using this now to see if our apex/functest CI is fixed
 > (tempest cases were failing).  But you are right if the fix is
 > merged we ignore it, but in the case where the code has not yet
 > been promoted into current-tripleo-rdo...we have no way of knowing
 > that.  I'm not sure what the easiest way to do that is.  Any ideas?

Checking in RDO repos if that commit has been included in a certain package is very tricky (we need mapping repo -> package name, check git log, etc). The easiest fix would be to ignore patches only if they are in ABANDONED/CLOSED state. In case the patch is "MERGED" we could try to apply the patch anyways, if it fails because the code is already there, catch the exception and just continue, without interrupting the deployment. Comments?

Patch-set: 1