Update patch set 19
Patch Set 19:

CI will only run the k8s when we run the gate.  The default CI will deploy openstack and make sure this patch doesn't break anything.  There was a bug upstream which was fixed and we were waiting for the patch to be promoted into the containers before this will pass.  We need a better mechanism of detecting when to pull a patch based on when the patch was promoted into the containers upstream.  I think Ricky has raised a JIRA for this already.

Patch-set: 19
Label: Verified=0