Update patch set 1
Patch Set 1:

> > Yeah just using this now to see if our apex/functest CI is fixed
 > > (tempest cases were failing).  But you are right if the fix is
 > > merged we ignore it, but in the case where the code has not yet
 > > been promoted into current-tripleo-rdo...we have no way of
 > knowing
 > > that.  I'm not sure what the easiest way to do that is.  Any
 > ideas?
 > 
 > Checking in RDO repos if that commit has been included in a certain
 > package is very tricky (we need mapping repo -> package name, check
 > git log, etc). The easiest fix would be to ignore patches only if
 > they are in ABANDONED/CLOSED state. In case the patch is "MERGED"
 > we could try to apply the patch anyways, if it fails because the
 > code is already there, catch the exception and just continue,
 > without interrupting the deployment. Comments?

Yeah that might work.  The only problem is if it is a python change it gets a little more complicated because we run the container updates in ansible in a loop.  So if one of the containers fails to patch, I think at that point we dont know if itw as a merged patch or an in progress patch.  So we don't know whether or not it is a real failure.

I had an idea that what we can do is if Merged, we check the time it was merged.  Then if it is a python patch, we check what time the last container was promoted.  Otherwise we assume it is a overcloud image patch, and we check what time the last overcloud image was pushed.  That should be fairly easy to add. I'll work on that.

Patch-set: 1
Label: Verified=0