Ignore error if network already undefined
In clean we destroy a network, then undefine it. When running snapshot
deploy it uses manual virsh bash commands to create the networks and
other virsh resources. This leaves an admin network which our clean
eventually tries to delete. However when the libvirt api network destroy
command is used it is somehow undefining the newtork at the same time.
Therefore just catch that possible scenario and ignore the failure as
the network has already been removed.

Change-Id: Id1d861fbe2a338f0828e78721c09f110030d3a4a
Signed-off-by: Tim Rozet <trozet@redhat.com>