Fix OpenDaylight Oxygen deployment failures (master)
ODL scenarios are failing deployment with oxygen because table 17 from
the OpenFlow pipeline is missing. This pulls in an upstream patch to no
longer check for table 17. The default table 17 flow has been removed
upstream which is why it no longer appears during deployment.

This patch fixes master OS branch scenarios only. Once the patch is
cherry-picked to queens we can submit a new patch updating the DS
settings.

Change-Id: If60cf25c75440e3a1998b412ac8b903849507dab
Signed-off-by: Tim Rozet <trozet@redhat.com>