Fixes for snapshots
With deploying snapshots with a new ODL, we currently bring down the
docker container and bring up the tar.gz distro of ODL on the Overcloud
host itself (not rebuilding/using container). Therefore we need java
installed so that ODL can run on the host. In the future this may
change, but it works well and keeps things simple for now.

Additionally, there was a change upstream to make the opendaylight
container docker restart policy "unless-stopped" which means it will
no longer restart automatically when docker is stopped/started.
Therefore on first snapshot bring up (without the previously mentioned
ODL reinstallation) the container does not start, and snapshot
deployment fails. This patch includes a change to the restart policy to
always restart it.

Change-Id: Icc712ba147e578a28e371313154ae3190676f0dc
Signed-off-by: Tim Rozet <trozet@redhat.com>