Update patch set 2
Patch Set 2:

Sure: One of the targets for the upcoming release is to support aarch64 on apex. From the container perspective it appears that we had to build 2 things for aarch64.
1) new kolla containers which can be seen here: https://registry.hub.docker.com/v2/repositories/armbandapex/ and as a result we bypass the ones build by tripleo (which are only for x86) 

2) Using a different ceph container (for aarch64) which can be found https://hub.docker.com/r/ceph/daemon/tags/ (master-fafda7d-luminous-centos-7-aarch64 currently)

So my only "issue" with this patch is that it does not leave room for me to bypass the x86 tag for ceph.

Patch-set: 2
Label: Verified=0