Update OVS in overcloud
This bug fix is not present in the RDO versions of OVS:
https://bugzilla.redhat.com/show_bug.cgi?id=1544892

The impact of this is that traffic sent through a patch port (like from
br-int to br-ex) will not be sent correctly with OF groups (which ODL
uses). This would cause tenant vlan traffic not to work in our case.

Change-Id: Ia477b174118dcd2d7fb24189c2d093218eb97053
Signed-off-by: Tim Rozet <trozet@redhat.com>