Allow common patches file
This patch adds allowing for common patches that should be applied to
every scenario to be included. It by default pulls in a file in the
deploy directory 'common-patches.yaml', but can optionally be
overridden.

This patch also includes a patch upstream to fix OSCLI not working
anymore due to breakage with the Cinder version in the overcloudrc.

Change-Id: I97b9efb937deff07e085b9ef75b9799fb65bfc57
Signed-off-by: Tim Rozet <trozet@redhat.com>