Modify common patches schema to per branch
The common patch design doesn't really work across branches very well.
This patch makes it so the common patches are specified in the same
file, but on a per branch basis. Making it easier to manage common
patches per branch.

Also, includes a fix and adds test coverage to the prep_image method
where we were referencing 'undercloud_admin_ip' before assignment in a
case where we were not using ODL and had patches to overcloud.

Change-Id: I7672947afd826fdc0042361a0139e22d6d5dd864
Signed-off-by: Tim Rozet <trozet@redhat.com>