Fix usage of list method of VirtualBMCManager
JIRA: APEX-638

As it can be observed from github, list function returns a tuple instead of a list.
It has been refactored for a long time.

https://github.com/openstack/virtualbmc/blob/master/virtualbmc/manager.py#L316

As the intention of the code is to access the list, this patch uses the second
item of tupple instead of the tupple itself.

Change-Id: I7af763008b69247b72aac22208b98061ca3736a8
Signed-off-by: Stamatis Katsaounis <mokats@intracom-telecom.com>