Update delorean repo in undercloud images
There is a new issue where the delorean repo that exists in the
undercloud disk images we get from rdo images is either expired or the
wrong repo. For example, the repo that exists in the queens undercloud
image is actually for openstack master. Rather than relying on the
delorean repo given in the disk image, this patch updates it to the
latest current-tripleo-rdo for the respective branch.

Change-Id: Ieb21f03dd5394f6b527f9304e36a95e8d18e9bb9
Signed-off-by: Tim Rozet <trozet@redhat.com>