Fixes failure to restart containers post undercloud install
It looks like the docker_container ansible module will recreate the
container if it fails to restart it. This is undesired behavior so
moving to use shell to restart the containers.

Also, fixes mistral executor container not properly mounting the
ceph-ansible playbook. Additionally fixes an issue with ceph-ansible by
downgrading the package. Related rhbz:
https://bugzilla.redhat.com/show_bug.cgi?id=1644713

Change-Id: I3199b4af11a4170d19419f70cb53f7d74def273c
Signed-off-by: Tim Rozet <trozet@redhat.com>