Fix Undercloud URL path
Modify the URL path for undercloud disk image to opnfv artifacts. The
undercloud image was removed from upstream TripleO, so we need to use
the latest known good version for Gambia release, stored in OPNFV
artifacts. In the future we will remove downloading undercloud image and
just use the overcloud image for undercloud.

Change-Id: I17a3bbaf67d55de05c195712d7fee6e89676bc16
Signed-off-by: Tim Rozet <trozet@redhat.com>