Update patch set 17
Patch Set 17:

> exactly my point.
 > So the if statements should branch be when --no-fetch has not been
 > used.

So we copy the images from the cache (unmodified), so we will still potentially need to expand the disk even if it is cached. I think the code checks if the disk is less than 50gb then it expands it to 50, so if you already had a big enough disk it wouldn't do anything. Can you give an example of where a user would use no-fetch and we dont want to expand the disk?

Patch-set: 17