Enable CSIT deployments to use NFS backend
In order to support HA as well as live migration a shared storage
backend needs to be used. This patch preps the overcloud image to start
with an NFS server sharing out of /root/nfs. If csit environment file is
used then NFS parameters will be passed to the deployment, configuring
Nova and Glance to use NFS hosted by controller0.

JIRA: APEX-654

Change-Id: I049d84bfd454013fc7ba5048939ee17100197624
Signed-off-by: Tim Rozet <trozet@redhat.com>