Fix NFS issues with Nova
There are problems with Nova launching instantces due to permissions
with nova being able to read/write certain directories on the NFS. The
permissions are right on the NFS and the folders the NFS mounts to, but
there still seems to be issues. The cause may be using a directory under
/root as the NFS mount. This patch moves the NFS mounts to be individual
folders under /. The patch also restart nova_compute docker container as
NFS problems still persist unless this is done.

JIRA: APEX-654

Change-Id: I25eee98c1a6516dfa44c686c2e614f6dc7000d98
Signed-off-by: Tim Rozet <trozet@redhat.com>