Always update OVS to 2.9.2
Right now RDO defaults to an old version of OVS. In Apex for non-sfc
scenarios we were pulling a prebuilt RPM of OVS 2.9.0 that included a
group tables fix needed for ODL. For SFC, we now just build OVS along
with the kernel module from native OVS because NSH support exists
upstream in that version.

We can move to always doing this operation for all OVS based scenarios,
which will bump our OVS and ensure the package we build at runtime will
work with our kernel.

Change-Id: If38d0b858878f1af66990d987612a8f8f6fa2718
Signed-off-by: Tim Rozet <trozet@redhat.com>