Renders service net map for THT
Previously if a service mapped to a network which was disabled, the
disabled network would fallback to using the ctlplane network. With the
recent change in THT, this is no longer the case:
https://review.openstack.org/#/c/614457/

With the above change, now any service pointing to a disabled network
now results in an empty string value being given for network variables.

This patch sets the service netmap in network-enviornment.yaml
appropriately based on which networks are enabled in apex.

Change-Id: Idf2919935aa707da6ca48968a04cf6653923d19d
Signed-off-by: Tim Rozet <trozet@redhat.com>