Fixes broken compute role update
We now insert the External network into the compute role after it was
removed upstream. However, the format has now changed for the network
specification. It no longer uses an Array, but instead uses a Dict. This
patch accounts for that case.

Also, adds new required arg --role-name to the NIC template merge tool.

Change-Id: Id3e4ecdfb1633ec4c58435c294f544a9625a106e
Signed-off-by: Tim Rozet <trozet@redhat.com>